PR cleanup
