--- experiments/nlp_experiments/classification/__init__.py
@@ -1,1 +1,0 @@
-from .csv_data import CSVClassificationDataset
