+++ setup.py
@@ -64,6 +64,6 @@ setup(
     extras_require={
         'test': ['pytest', 'pytest-pep8', 'hypothesis', 'coverage', 'coveralls'],
         'documentation': documentation_requirements,
-        'dev': ['transformers', 'sentencepiece']
+        'dev': ['transformers', 'datasets']
     },
 )
